+++ CHANGELOG.md
@@ -22,6 +22,27 @@ All notable changes to this project will be documented here. For more details, v
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## UNRELEASED
+
+### Added
+
+- 1fichier support
+- Archive.org support
+- YTboob support
+- Rutube support
+
+### Changed
+
+### Deprecated
+
+### Removed
+
+### Fixed
+
+- Date extraction (Rule34vault)
+
+### Security
+
 ## [9.11.0] - 2026-05-20
 
 ### Changed